экран информации о рабочем месте

+++ app/src/main/java/com/example/healthyapp/db/model/entity/WorkplaceUser.kt
@@ -2,11 +2,11 @@ package com.example.healthyapp.db.model.entity
 
 data class WorkplaceUser(
     val id: String,
-    val height: Int,
-    val sitHeight: Int,
-    val legsHeight: Int,
-    val eyesHeight: Int,
-    val shoulder: Int,
-    val back: Int,
+    val height: Long,
+    val sitHeight: Long,
+    val legsHeight: Long,
+    val eyesHeight: Long,
+    val shoulder: Long,
+    val back: Long,
     val name: String
 )
